Refactor data classes to reflect refactor in start
