+++ src/main/java/com/example/application/data/CompanyRepository.java
@@ -1,6 +1,6 @@
-package com.example.application.data.repository;
+package com.example.application.data;
 
-import com.example.application.data.entity.Company;
+import com.example.application.data.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CompanyRepository extends JpaRepository<Company, Long> {
