+++ src/main/java/com/example/application/data/ContactRepository.java
@@ -1,6 +1,6 @@
-package com.example.application.data.repository;
+package com.example.application.data;
 
-import com.example.application.data.entity.Contact;
+import com.example.application.data.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
