+++ src/main/java/com/example/application/data/StatusRepository.java
@@ -1,6 +1,6 @@
-package com.example.application.data.repository;
+package com.example.application.data;
 
-import com.example.application.data.entity.Status;
+import com.example.application.data.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StatusRepository extends JpaRepository<Status, Long> {
