+++ src/main/java/com/example/application/views/list/ListView.java
@@ -1,7 +1,7 @@
 package com.example.application.views.list;
 
-import com.example.application.data.entity.Contact;
-import com.example.application.data.service.CrmService;
+import com.example.application.data.Contact;
+import com.example.application.services.CrmService;
 import com.example.application.views.MainLayout;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.button.Button;
